for review

+++ Services/Tasks/src/Domain.Tasks/Entities/CountPrimes/CountPrimesTaskEntity.cs
@@ -1,9 +1,8 @@
 ﻿using Domain.Tasks.Abstracts.Existing;
-using Domain.Tasks.Entities.Factorial;
 
 namespace Domain.Tasks.Entities.CountPrimes;
 
-public class CountPrimesTaskEntity : TaskEntityBase<CountPrimesTaskArtefacts, int, long>
+public class CountPrimesTaskEntity : TaskEntityBase<CountPrimesTaskArtefacts, int, int>
 {
     public override required CountPrimesTaskArtefacts Artefacts { get; set; }
 }
